Add optional config option to delay sending kill signal
This adds a new `delay` option to the [stop-button] section. This specifies the delay in seconds between the time when the button is pressed and when the signal is sent to the remote processes.

Releasing the button before the signal is sent will cancel sending the signal.

Issue: ev3dev/ev3dev#1162